requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic-settings==2.11.0
openai==1.108.2
faiss-cpu==1.12.0
numpy==2.3.3
pypdf==6.1.1
python-docx==1.2.0
pytest==8.4.2
httpx==0.28.1
